Develop

+++ app-server/src/main/java/com/inventory/app/server/service/media/GameService.java
@@ -1,8 +1,11 @@
 package com.inventory.app.server.service.media;
 
 import com.inventory.app.server.entity.media.Game;
+import com.inventory.app.server.error.NoChangesToUpdateException;
+import com.inventory.app.server.error.ResourceAlreadyExistsException;
+import com.inventory.app.server.error.ResourceNotFoundException;
 import com.inventory.app.server.repository.IBaseDao;
-import com.inventory.app.server.utility.RestPreConditions;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,10 @@ import java.util.List;
 public class GameService {
     private IBaseDao<Game> dao;
 
+    public GameService(IBaseDao<Game> dao) {
+        this.dao = dao;
+    }
+
     @Autowired
     public void setDao(IBaseDao<Game> daoToSet) {
         dao = daoToSet;
@@ -20,59 +27,95 @@ public class GameService {
     }
 
     public List<Game> getAllGamesByNumberOfPlayers(Integer numberOfPlayers){
-        return dao.findByField("number_of_players", String.valueOf(numberOfPlayers));
+        List<Game> gameList = dao.findByField("number_of_players", String.valueOf(numberOfPlayers));
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No games found with number of players: " + numberOfPlayers);
+        }
+        return gameList;
     }
 
     public List<Game> getAllGamesByConsole(List<String> console){
-        return  dao.findByField("consoles", console);
+        List<Game> gameList = dao.findByField("consoles", console);
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No games found for consoles " + console );
+        }
+        return gameList;
     }
 
     public List<Game> getAllGamesByTitle(String title){
-        return dao.findByField("title", title);
+        List<Game> gameList = dao.findByField("title", title);
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No game results for title " + title);
+        }
+        return gameList;
     }
 
     public List<Game> getAllGamesByCollectionTitle(String collectionTitle) {
-        return dao.findByField("collection_name", collectionTitle);
+        List<Game> gameList = dao.findByField("collection_name", collectionTitle);
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No game results found for collection title " + collectionTitle);
+        }
+        return gameList;
     }
 
     public List<Game> getAllGamesByGenre(String genre) {
-        return dao.findByField("genre", genre);
+        List<Game> gameList = dao.findByField("genre", genre);
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No fame results found for genre " + genre);
+        }
+        return gameList;
     }
 
-    public List<Game> getAllGames() {
-        return dao.findAll();
+    public List<Game> getAll() {
+        List<Game> gameList = dao.findAll();
+        if (gameList.isEmpty()) {
+            throw new ResourceNotFoundException("No game data exists.");
+        }
+        return gameList;
     }
 
-    public Game getGameById(Long id) {
-        return dao.findOne(id);
+    public Game getById(Long id) {
+        try {
+            return dao.findOne(id);
+        } catch (Exception e) {
+            if(e.getClass().isInstance(EntityNotFoundException.class)) {
+                throw new ResourceNotFoundException("No game exists with id: " + id);
+            } else {
+                throw e;
+            }
+        }
     }
 
     public Game create(Game game) {
-        // validations before performing create
-        RestPreConditions.checkAlreadyExists(gameAlreadyExists(game), game);
-
+        if (gameAlreadyExists(game)) {
+            throw new ResourceAlreadyExistsException("Cannot create game because games already exist: " + game);
+        }
         Game gameToSave = cloneGame(game);
         gameToSave.setId(null);
         gameToSave.setVersion(1);
-
         return dao.createOrUpdate(gameToSave);
     }
 
     public Game update(Game updatedGame) {
-        //validations before performing the update
-        Game existingGame = RestPreConditions.checkFound(getGameById(updatedGame.getId()));
-        RestPreConditions.checkEquals(existingGame, updatedGame);
-
+        if (!gameAlreadyExists(updatedGame)) {
+            throw new ResourceNotFoundException("Cannot update game because game does not exist: " + updatedGame);
+        }
+        Game existingGame = getById(updatedGame.getId());
+        if (verifyIfGameUpdate(existingGame, updatedGame)) {
+            throw new NoChangesToUpdateException("No updates in book to save. Will not proceed with update. Existing Game: " + existingGame + " Update Game: " + updatedGame);
+        }
         updatedGame = cloneGame(existingGame, updatedGame);
         updatedGame.setVersion(existingGame.getVersion() + 1);
-
         return dao.createOrUpdate(updatedGame);
     }
 
     public Game deleteById(Long id){
-       Game game = RestPreConditions.checkFound(getGameById(id));
-        dao.deleteById(id);
-        return game;
+       Game game = getById(id);
+       if (game == null) {
+           throw new ResourceNotFoundException("Cannot delete game because game does not exist.");
+       }
+       dao.deleteById(id);
+       return game;
     }
 
     private Game cloneGame(Game game) {
@@ -90,5 +133,9 @@ public class GameService {
         return getAllGamesByTitle(game.getTitle())
                 .stream()
                 .anyMatch(g -> game.getTitle().equals(g.getTitle()));
+    }
+
+    private boolean verifyIfGameUpdate(Game existingGame, Game updatedGame) {
+        return existingGame.equals(updatedGame);
     }
 }
